src/main/java/com/skb/TaskManagementService/service/TaskServiceImpl.java
package com.skb.TaskManagementService.service;

import com.skb.TaskManagementService.model.Task;
import com.skb.TaskManagementService.repository.TaskRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class TaskServiceImpl implements TaskService {

    @Autowired
    private TaskRepository taskRepository;

    @Override
    public List<Task> getAllTasks() {
        return taskRepository.findAll();
    }

    @Override
    public Task createTask(Task task) {
        return taskRepository.save(task);
    }

    @Override
    public Task updateTask(Long taskId, String status) {
        Task task = taskRepository.findById(taskId).orElse(null);
        if (task != null) {
            task.setStatus(status);
            return taskRepository.save(task);
        }
        return null;
    }

    @Override
    public boolean deleteTask(Long taskId) {
        if (taskRepository.existsById(taskId)) {
            taskRepository.deleteById(taskId);
            return true;
        }
        return false;
    }

    @Override
    public Task assignTask(Long taskId, String userId) {
        Task task = taskRepository.findById(taskId).orElse(null);
        if (task != null) {
            task.setAssignedTo(userId);
            return taskRepository.save(task);
        }
        return null;
    }
}